Refactor Geom into MeshTools
